refactor(main): 데이터 입력 클래스 분리
- 데이터 입력, 입력데이터 유효성 검사, 입력데이터 배열로 변경 기능을 다른 클래스로 분리
